# src/main/resources/META-INF/openapi.yml
openapi: 3.0.2
info:
  title: 'OpenAPI spec for Raspberry Pi #growlab resources'
  version: 1.0.0
  description: |-
    OpenAPI spec for Raspberry Pi #growlab resources:
      - GET temperature, humidity, and pressure readings
      - GET light intensity readings
      - GET RPi Camera module image
      - GET RPi Camera module current mode
      - PUT to set RPi Camera module current mode
      - GET Health checks
servers:
  - url: 'http://192.168.1.73:8088/growlab/api/v1'
    description: dev
  - url: 'http://192.168.1.72:8080/growlab/api/v1'
    description: home lab
paths:
  /actuators/camera/image:
    summary: Take and get an image from the Raspberry Pi camera
    description: This operation allows to take and get an image from the Raspberry Pi camera.
    get:
      responses:
        '200':
          $ref: '#/components/responses/image_response'
        '500':
          $ref: '#/components/responses/actuator_error'
      operationId: getImage
      summary: Take and get an image from the Raspberry Pi camera
      description: This operation allows to take and get an image from the Raspberry Pi camera
  /sensors/light:
    summary: The light intensity sensor readings
    description: |-
      This resource allows to obtain light intensity readings
      from the light intensity sensor, e.g. BH1750
    get:
      tags: [ ]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sensor_readings_item_type'
              examples:
                lignt_intensity_readings_example1:
                  value:
                    s_type: LIGHT
                    s_data: '1.67'
                    s_timestamp: '2021-08-22T22:44:20.369467Z'
          description: The light intensity sensor readings
        '500':
          $ref: '#/components/responses/sensor_error'
      operationId: getLightIntensity
      summary: Get light intensity readings
      description: Can be used to get readings the light intensity sensor
  /actuators/camera/mode:
    summary: The Raspberry Pi IR-CUT camera mode control
    description: This resource can be used to set or get the Raspberry Pi IR-CUT camera operation mode
    get:
      responses:
        '200':
          $ref: '#/components/responses/camera_mode_response'
        '500':
          $ref: '#/components/responses/actuator_error'
      operationId: getCameraMode
      summary: Get camera mode
      description: Can be used to get the camera current mode of operation
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/camera_mode_set_type'
            examples:
              set_camera_mode_example_1:
                value:
                  mode: NORM
        required: true
      responses:
        '200':
          description: The camera mode update successfully
        '500':
          $ref: '#/components/responses/actuator_error'
      operationId: putCameraMode
      summary: Set the camera operation mode
      description: Can be used to set the camera operation mode
  /health:
    summary: Provides health information of the runtime
    description: |-
      Provides the readiness and the liveness of the runtime according to
      the Eclipse MicroProfile Health Check specification.
      Please, see https://github.com/eclipse/microprofile-health for further details
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: https://raw.githubusercontent.com/eclipse/microprofile-health/3.1/tck/src/main/resources/health-check-schema.json
              examples:
                health_check_example1:
                  value:
                    status: UP
                    checks:
                      - name: deadlock
                        status: UP
                      - name: diskSpace
                        status: UP
                        data:
                          free: 3.92 GB
                          freeBytes: 4208271360
                          percentFree: 56.61%
                          total: 6.92 GB
                          totalBytes: 7434330112
                      - name: heapMemory
                        status: UP
                        data:
                          free: 21.38 MB
                          freeBytes: 22421208
                          max: 61.88 MB
                          maxBytes: 64880640
                          percentFree: 84.36%
                          total: 31.06 MB
                          totalBytes: 32571392
                      - name: systemInfo
                        status: UP
                        data:
                          cpuTemp: '37.932'
          description: The runtime health information
        '500':
          $ref: '#/components/responses/sensor_error'
      operationId: getHealthChecks
      summary: Can be used to get health checks
      description: Can be used to get the health information of the runtime
  /sensors/thp:
    summary: 'The temperature, humidity and pressure sensor readings'
    description: |-
      This resource allows to obtain temperature, humidity, and pressure readings from the
      environmental data sensor, e.g. BME280
    get:
      tags: [ ]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sensor_readings_type'
              examples:
                thp_readings_example:
                  value:
                    s_readings:
                      - s_id: '60'
                        s_type: TEMP
                        s_data: '26.079965189471842'
                        s_timestamp: '2021-08-22T18:10:55.736630Z'
                      - s_id: '60'
                        s_type: HUMID
                        s_data: '45.481067387147625'
                        s_timestamp: '2021-08-22T18:10:55.736630Z'
                      - s_id: '60'
                        s_type: PRESS
                        s_data: '99556.74677629767'
                        s_timestamp: '2021-08-22T18:10:55.736630Z'
          description: |-
            The collection of temperature, humidity, and pressure readings, obtained from
            the environmental data sensor, e.g. BME280
        '500':
          $ref: '#/components/responses/sensor_error'
      operationId: getThp
      summary: 'Get temperature, humidity and pressure readings'
      description: |-
        Can be used to get readings of temperature, humidity, and pressure readings,
        obtained from the environmental data sensor, e.g. BME280
components:
  schemas:
    sensor_readings_type:
      description: An array of sensors readings
      type: object
      properties:
        s_readings:
          description: Sensor readings
          type: array
          items:
            $ref: '#/components/schemas/sensor_readings_item_type'
      example:
        s_readings:
          - s_id: '60'
            s_type: TEMP
            s_data: '26.079965189471842'
            s_timestamp: '2021-08-22T18:10:55.736630Z'
          - s_id: '60'
            s_type: HUMID
            s_data: '45.481067387147625'
            s_timestamp: '2021-08-22T18:10:55.736630Z'
          - s_id: '60'
            s_type: PRESS
            s_data: '99556.74677629767'
            s_timestamp: '2021-08-22T18:10:55.736630Z'
    sensor_readings_item_type:
      description: 'A structure, containing sensor data readings, e.g. temperature, humidity, pressure, etc.'
      required:
        - s_data
        - s_timestamp
      type: object
      properties:
        s_type:
          description: 'A sensor type, e.g. TEMP (temperature)'
          type: string
        s_id:
          description: Sensor id
          type: string
        s_data:
          description: A sensor readings value
          type: string
        s_timestamp:
          format: date-time
          description: Readings fetch timestamp
          type: string
      example:
        s_id: '60'
        s_type: TEMP
        s_data: '26.079965189471842'
        s_timestamp: '2021-08-22T18:10:55.736630Z'
    sensor_error_type:
      description: This type contains a sensor readings request error description elements of the API
      required:
        - event_id
        - event_name
        - event_timestamp
      type: object
      properties:
        event_id:
          description: An error event type id
          type: string
        event_name:
          description: An error event descriptive name
          type: string
        s_type:
          description: 'A sensor type, e.g. TEMP (temperature)'
          type: string
        desc:
          description: An error event detailed info
          type: string
        event_timestamp:
          format: date-time
          description: An error event occurrence timestamp
          type: string
      example:
        event_id: E-BH1750-0001
        event_name: BH1750 data read error
        s_type: LIGHT
        desc: >-
          dev.sergevas.iot.growlabv1.shared.exception.SensorException: BH1750 data read errorCaused
          by: java.lang.RuntimeException: Test exception
        event_timestamp: '2021-08-22T20:34:51.809518Z'
    actuator_error_type:
      description: This type contains an actuator command request error description elements of the API
      required:
        - event_id
        - event_name
        - event_timestamp
      type: object
      properties:
        event_id:
          description: An error event type id
          type: string
        event_name:
          description: An error event descriptive name
          type: string
        desc:
          description: An error event detailed info
          type: string
        event_timestamp:
          format: date-time
          description: An error event occurrence timestamp
          type: string
      example:
        event_id: E_CAMERA_0001
        event_name: Camera image take error
        desc: 'dev.sergevas.iot.growlabv1.shared.application.port.out.ActuatorException: Camera image take error'
        event_timestamp: '2021-08-22T20:34:51.809518Z'
    camera_mode_type:
      title: Root Type for camera_mode_type
      description: The Raspberry Pi IR-CUT camera current mode of operation.
      required:
        - mode
        - mode_timestamp
      type: object
      properties:
        mode:
          description: The Raspberry Pi IR-CUT camera mode.
          enum:
            - NORM
            - NIGHT
          type: string
        mode_timestamp:
          format: date-time
          description: The camera mode fetch timestamp
          type: string
      example:
        mode: NIGHT
        mode_timestamp: '2021-08-22T16:02:11.452302Z'
    camera_mode_set_type:
      title: Root Type for camera_mode_type
      description: The Raspberry Pi IR-CUT camera operation mode.
      required:
        - mode
      type: object
      properties:
        mode:
          description: |-
            The Raspberry Pi IR-CUT camera operation mode.
            Possible values:
             - NORM - normal mode
             - NIGHT - night mode
          enum:
            - NORM
            - NIGHT
          type: string
      example:
        mode: NORM
  responses:
    image_response:
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
      description: A camera inage binary content
    sensor_error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/sensor_error_type'
          examples:
            sensor_error_example_1:
              value:
                event_id: E-BH1750-0001
                event_name: BH1750 data read error
                s_type: LIGHT
                desc: >-
                  dev.sergevas.iot.growlabv1.shared.exception.SensorException: BH1750 data
                  read errorCaused by: java.lang.RuntimeException: Test exception
                event_timestamp: '2021-08-22T20:34:51.809518Z'
      description: A sensor error message
    actuator_error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/actuator_error_type'
          examples:
            actuator_error_example_1:
              value:
                event_id: E_CAMERA_0001
                event_name: Camera image take error
                desc: >-
                  dev.sergevas.iot.growlabv1.shared.exception.ActuatorException: Camera
                  image take error
                event_timestamp: '2021-08-22T20:34:51.809518Z'
            actuator_error_example_2:
              value:
                event_id: E-CAMERA-0003
                event_name: Camera mode update error
                desc: >-
                  dev.sergevas.iot.growlabv1.shared.exception.ActuatorException: Camera mode
                  update error
                event_timestamp: '2021-08-22T20:34:51.809518Z'
            actuator_error_example_3:
              value:
                event_id: E-CAMERA-0002
                event_name: Camera mode read error
                desc: >-
                  dev.sergevas.iot.growlabv1.shared.exception.ActuatorException: Camera mode
                  read error
                event_timestamp: '2021-08-22T20:34:51.809518Z'
      description: An actuator error message
    camera_mode_response:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/camera_mode_type'
      description: The Raspberry Pi IR-CUT camera current mode of operation.